requirements.txt
# Core dependencies
python-telegram-bot==13.15  # Telegram API wrapper
pytz==2025.2                # Timezone handling
tzlocal==5.3.1              # Local timezone detection
APScheduler==3.6.3          # Job scheduling (compatible with PTB 13.x)

# Web & Networking
urllib3==1.26.18            # HTTP client (for backward compatibility)
certifi>=2023.7.22          # SSL certificates

# Data handling & formatting
tabulate==0.9.0             # Pretty table formatting for admin dashboard
setuptools>=42.0.0          # Required by APScheduler
typing-extensions>=4.0.0    # Enhanced type hints support

# Utils
six>=1.10.0                 # Python 2/3 compatibility (required by some dependencies) 